--- Quackademics/src/components/OverallStudyRanking.jsx
@@ -1,23 +1,0 @@
-import * as React from 'react';
-import Box from '@mui/material/Box';
-
-const OverrallStudyRanking = () => {
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        backgroundColor: "#615f5f",
-        color: "white",
-        borderRadius: 2,
-        width: "100%",
-        height: "100%"
-      }}
-    >
-      Overrall Study Rankings Board
-    </Box>
-  )
-}
-
-export default OverrallStudyRanking;
