--- Quackademics/src/components/StatsGraph.jsx
@@ -1,25 +1,0 @@
-import * as React from 'react';
-import Box from '@mui/material/Box';
-
-const StatsGraph = () => {
-  return (
-    <>
-      <Box
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          backgroundColor: "#615f5f",
-          color: "white",
-          borderRadius: 2,
-          height: "45%",
-          width: "100%",
-        }}
-      >
-        Stats Graph 
-      </Box>
-    </>
-  )
-}
-
-export default StatsGraph;
